Code architecture refactoring
Reimplement the code in a modular way.
